refactor/ zustand store.ts 파일 분리 작업 - 기존에 모든 전역 상태를 store.ts 파일에 모아 두었고, 이를 각 상태의 그룹 별로 파일을 분리하는 작업 수행
