+++ src/components/UI/styler/styler/text-styler/TextStrokeThicknessStyler.tsx
@@ -1,4 +1,4 @@
-import { useQuotesStrokeStyleStore } from "@/store/store"
+import { useQuotesStrokeStyleStore } from "@/store/stylerStore"
 import { debounceCloser } from "@/utils/common-func"
 import { ChangeEvent, useRef } from "react"
 import { HiMinus, HiPlus } from "react-icons/hi2"
@@ -40,16 +40,16 @@ export default function TextStrokeThicknessStyler() {
     return (
 
         <article className='max-w-[150px] px-[10px] rounded-sm flex items-center relative ml-[0.5em] border '>
-                <button className='text-white hover:text-[#ffcdc5]  text-[1.15em] pr-[5px]' onClick={decrease}><HiMinus /></button>
-                <input
-                    value={DEFAULT_TEXT_SIZE}
-                    min={0}
-                    ref={inputRef}
-                    type="number"
-                    className="p-[15px] text-center rounded-[3px] focus:outline-none bg-transparent text-white h-[35px] max-w-[100px] w-full"
-                    onChange={onSetThickness}
-                />
-                <button className='text-white hover:text-[#97ef97] text-[1.15em] pl-[5px]' onClick={increase}><HiPlus /></button>
+            <button className='text-white hover:text-[#ffcdc5]  text-[1.15em] pr-[5px]' onClick={decrease}><HiMinus /></button>
+            <input
+                value={DEFAULT_TEXT_SIZE}
+                min={0}
+                ref={inputRef}
+                type="number"
+                className="p-[15px] text-center rounded-[3px] focus:outline-none bg-transparent text-white h-[35px] max-w-[100px] w-full"
+                onChange={onSetThickness}
+            />
+            <button className='text-white hover:text-[#97ef97] text-[1.15em] pl-[5px]' onClick={increase}><HiPlus /></button>
         </article>
     )
 }
